+++ packages/core-cairo/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Support custom contract type, optional access control. ([#140](https://github.com/OpenZeppelin/contracts-wizard/pull/140))
+
 ## 0.2.0 (2022-06-15)
 
 - Update API format for Cairo. ([#136](https://github.com/OpenZeppelin/contracts-wizard/pull/136))